+++ extensions/apple-notes/src/new.tsx
@@ -1,6 +1,5 @@
 import { closeMainWindow, getSelectedText, LaunchProps, showToast, Toast } from "@raycast/api";
-import { runAppleScript } from "@raycast/utils";
-import { escapeDoubleQuotes } from "./utils";
+import { createNote } from "./api";
 
 export default async (props: LaunchProps<{ arguments: Arguments.New }>) => {
   await closeMainWindow();
@@ -21,24 +20,7 @@ export default async (props: LaunchProps<{ arguments: Arguments.New }>) => {
   }
 
   try {
-    const script = text.trim()
-      ? `
-    set noteContent to "${escapeDoubleQuotes(text)}"
-    tell application "Notes"
-      activate
-      set newNote to make new note at folder "Notes"
-      set body of newNote to noteContent
-      set selection to newNote
-    end tell
-    `
-      : `
-    tell application "Notes"
-      activate
-      make new note at folder "Notes"
-    end tell
-    `;
-
-    await runAppleScript(script);
+    await createNote(text.trim());
   } catch (error) {
     showToast({ style: Toast.Style.Failure, title: "Could not create a new note." });
   }