+++ extensions/apple-notes/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Improvements] - 2024-03-29
 
 - Add locked accessory
+- Empty state
 
 ## [Better search] - 2024-03-22